+++ apps/app-chat/src/main/java/vip/qsos/app_chat/data/entity/ChatMessageBo.kt
@@ -29,7 +29,7 @@ data class ChatMessageBo(
             message.messageId = value
         }
 
-    override var sessionId: Int = -1
+    override var sessionId: Long = -1L
         get() {
             field = message.sessionId
             return field
